### 2019-07-19 #325 @nfox29
Submission: photosearcher
Submitting Author: Name (@nfox29)
Repository: [nfox29/photosearcher](https://github.com/nfox29/photosearcher)
Version submitted: 1.0
Version accepted: TBD

---

```
Package: photosearcher
Title: Photo Searcher
Version: 1.0
    c(person("Nathan", "Fox", role = c("aut", "cre"), email = "nf2g13@soton.ac.uk"),
           person("Francesca", "Mancini", role = c("aut")),
           person("Laura", "Graham", role = c("aut")),
           person("Louis", "Sutter", role = c("aut")),
           person("Tom", "August", role = c("aut")))
Description: Queries the Flick API (https://www.flickr.com/services/api/)to return photograph metadata as well as the ability to download the images as jpegs.
License: GPL
Language: en_GB
    xml2,
    httr,
    glue,
    clisymbols,
    crayon,
    sf
    httptest,
    roxygen2,
    covr,
    rplos,
    wordcloud,
    tidyverse
```

The package provides functions that make calls to the [Flickr API](https://www.flickr.com/services/api/) to return metadata (including georeferenced location) for photographs in a reproducible manner.

The package is aimed to facilitate reproducible code for scientific research that utilises the Flickr API. Currently, a wide range of scientific disciplines utilise the Flickr API including environmental sciences, biological sciences and medicine. Flickrs georeferenced images have a wide range of scientific applications including ecosystem service assessments, resource management decisions and mapping species distributions.

We are aware of [one other package](http://www.maths.bristol.ac.uk/R/web/packages/FlickrAPI/index.html) that makes calls to the Flickr API. However this package only allows user to search for a single image based on its ID number. Our package provides users with the ability to search for photographs based on a user defined criteria including, date of the photograph being taken, tags, text and spatial location. Furthermore our package provides additional functionality including the downloading of images, retrieving social data about users as well as get tags related to a specific location, or other tags.
