### 2019-07-22 #325 @noamross
Thanks for your submission, @nfox29!  I have a question regarding the relationship of this package and **FlickrAPI**.  A quick review suggests to me that these packages cover different, but complementary parts of the API.  Is that correct, or is the difference in how API outputs are handled?

I ask. We generally aim for RO packages to be best-in-category because we want the RO package to be the one we'd recommend for the task. In this case, it seems the best outcome would be to combine the functionality of the two packages, either by joining forces, or you adding the parts of the API that **FlickrAPI** has, so that we have one best package for accessing Flickr.  Would you agree or do you think the packages handle very different applications?
