### 2019-07-22 #325 @nfox29
Thank you @noamross for your swift reply. I appreciate that there seems to be similarities between the two packages.

flickrAPI has four functions getExif, getHotTags, getPhotoInfo and getPhotos. Out of those four we currently do not have getExif, getHotTags and getPhotos.

The main function we share is searching for photo metadata. We use the API method [flickr.photos.search](https://www.flickr.com/services/api/flickr.photos.search.html) which allows users to search for photos based on their own arguments including tags and spatial location. flickrAPI on the other hand used the API method [flickr.photos.getInfo]( https://www.flickr.com/services/api/flickr.photos.getInfo.html). The difference with this method is that it only allows user to search for information on one specific image. The argument to search for a specific image by its ID can be incorporate into our method. Second the way that data is returned is handled slightly differently. The use of our method allows users to return all available metadata on an image to a single dataframe, whereas the call made by flickrAPI returns limited data for one of "Location", "Date", "URL", "Tags".

Of the functions that we do not have in our package we have chosen not to include the getHotTags function as this API method does not work Flickr side and will only ever return the same result regardless off the arguments. With the other two, flickrAPI doesn’t utilise the full range of arguments available to the calls (for example the getPhotos function in flickrAPI only has one augment, whereas the API method has many more). If you deem it necessary to include these functions I will add them to this package with the greater functionality available on the API.

Besides that – our package has additional API methods, including downloading images, getting the top tags for locations and assessing related tags. Our package also has some quality of life benefits, such as storing you API key so that you do not have to enter it as an argument to the functions every time.

I will now add the option to search for a photos data via its ID number to the photosearch function, as well as adding the two other functions.
