### 2019-07-23 9232bfd
- download images now skips images if they no longer exist
