### 2019-07-23 32c0282
- added get_exif function
