+++ man/get_exif.Rd
@@ -1,0 +1,23 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/get_exif.R
+\name{get_exif}
+\alias{get_exif}
+\title{get_exif}
+\usage{
+get_exif(photo_id = NULL)
+}
+\arguments{
+\item{photo_id}{Id of photograph}
+}
+\value{
+A dataframe of "exchangeable image file format" information for the
+given photograph
+}
+\description{
+get_exif
+}
+\examples{
+\dontrun{
+get_exif(photo_id = 123)
+}
+}
