### 2019-07-23 2943021
- photo_search has added agruements of user_id, mindate_uploaded and maxdate_uploaded.
- now correctly returns description field
