+++ R/photo_search.R
@@ -9,13 +9,18 @@
 #' \url{https://www.flickr.com/services/api/flickr.photos.search.html} for more
 #' information on the API method.
 #'
-#' @param mindate character. Minimum date of photograph for search provided as
-#'   "YYYY-MM-DD".
-#' @param maxdate character. Maximum date of photograph for search provided as
-#'   "YYYY-MM-DD".
-#'
-#' @param text String, optional text to be searched.
-#' @param tags String, optional tags to filter by.
+#' @param mindate_taken Character, required. Minimum taken date. Photos with an
+#'   taken date greater than or equal to this value will be returned. The date
+#'   should be in the form of "YYYY-MM-DD".
+#' @param maxdate_taken Character, required. Maximum taken date. Photos with an
+#'   taken date less than or equal to this value will be returned. The date
+#'   should be in the form of "YYYY-MM-DD".
+#' @param text String, optional. A free text search. Photos who's title,
+#'   description or tags contain the text will be returned. You can exclude
+#'   results that match a term by prepending it with a - character.
+#' @param tags String, optional. A comma-delimited list of tags. Photos with one
+#'   or more of the tags listed will be returned. You can exclude results that
+#'   match a term by prepending it with a - character.
 #' @param bbox String, optional bounding box of search area provide as:
 #'   "minimum_longitude,minimum_latitude,maximum_longitude,maximum_latitude".
 #' @param woe_id Numeric, optional "where on earth identifier" can be supplied
@@ -24,6 +29,15 @@
 #'   can be supplied instead of a bbox or woeID.
 #' @param has_geo Logical, optional argument for whether returned photos need
 #'   associated spatial data.
+#' @param mindate_uploaded Character, optional. Minimum upload date. Photos with
+#'   an upload date greater than or equal to this value will be returned. The
+#'   date can be in the form of a unix timestamp or mysql datetime.
+#' @param maxdate_uploaded Character, optional. Maximum upload date. Photos with
+#'   an upload date less than or equal to this value will be returned. The date
+#'   can be in the form of a unix timestamp or mysql datetime.
+#' @param user_id Character, optional. The NSID of the user who's photo to
+#'   search. If this parameter isn't passed then everybody's public photos will
+#'   be searched.
 #'
 #' @return data.frame. Output consists of 57 variables including; latitude and
 #'   longitude of photograph, date and time it was taken, associated tags and
@@ -44,8 +58,8 @@
 #' @examples
 #' \dontrun{
 #' photo_search(
-#'   mindate = "2019-01-01",
-#'   maxdate = "2019-01-02",
+#'   mindate_taken = "2019-01-01",
+#'   maxdate_taken = "2019-01-02",
 #'   text = "tree",
 #'   bbox = "-13.623047,47.279229,3.251953,60.630102",
 #'   has_geo = TRUE
@@ -53,23 +67,37 @@
 #'
 #'
 #' photo_search(
-#'   mindate = "2019-01-01",
-#'   maxdate = "2019-01-02",
+#'   mindate_taken = "2019-01-01",
+#'   maxdate_taken = "2019-01-02",
 #'   text = "tree",
 #'   bbox = "-13.623047,47.279229,3.251953,60.630102",
-#'   has_geo = TRUE
-#' )
+#'   has_geo = TRUE)
+#'
+#'   photo_search(
+#'    mindate_taken = "2017-01-01",
+#'    maxdate_taken = "2017-01-02",
+#'    mindate_uploaded = "2017-03-04",
+#'    maxdate_uploaded = "2017-05-05",
+#'    tags = "tree")
+#'
+#'    photo_search(
+#'     mindate_taken = "2014-01-01",
[…]
