### 2019-07-23 5efdfc8
- changed failing tests
