### 2019-07-23 6e45bff
- added function to search for single photos info
- added tests for get_exif and get_photoinfo
