### 2019-07-23 8b57e61
- updated failing tests
- updated vingette and docs
