+++ man/get_photoinfo.Rd
@@ -1,0 +1,23 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/get_photoinfo.R
+\name{get_photoinfo}
+\alias{get_photoinfo}
+\title{get_photoinfo}
+\usage{
+get_photoinfo(photo_id = NULL)
+}
+\arguments{
+\item{photo_id}{Character, required. The id of the photo to get information
+for.}
+}
+\value{
+Dataframe of information on given image
+}
+\description{
+get_photoinfo
+}
+\examples{
+\dontrun{
+get_photoinfo(photo_id = 123)
+}
+}
