+++ NAMESPACE
@@ -3,6 +3,7 @@
 export(download_images)
 export(find_place)
 export(get_exif)
+export(get_photoinfo)
 export(location_tags)
 export(photo_search)
 export(related_terms)
