### 2019-07-23 56e599b
- added package qualifiers to functions
