### 2019-07-23 #325 @nfox29
Hi @noamross, to address the issues you raised I have now updated the package to include the functionality that **flickrAPI** included that we originally did not.  Our package can now:

- search for a data on a single image via the `get_photoinfo` function
- search for exif data for an image via the `get_exif` function
- search for the photographs uploaded by a single user under the `photo_search` function

I have also updated testing, docs and vingettes to match these updates.

I believe our package now covers the complete functionality of **flickrAPI**, whilst providing additional functions and returning the data in a more complete way. I hope these resolves the issue and the package can be considered for submission.
Kindest regards,
Nathan
