### 2019-07-29 #325 @sckott
### Editor checks:

---

#### Editor comments

Thanks for your submission @nfox29      !

License: you have `GPL` in your license file. I'm not sure I've seen that before. I usually see GPL-2, GPL-3, etc. - see https://cran.r-project.org/doc/manuals/r-release/R-exts.html#Licensing - Did you mean `GPL` per se?

Here's the output from [goodpractice](https://github.com/MangoTheCat/goodpractice/). If you haven't used `goodpractice` it's an R package that checks a number of things with another package - most of which we agree with and want authors to follow. You don't need to address these now, it's info for reviewers to use to get started.

- There's some code lines over 80 characters in width, but most are within 80, so good job on that front - good to get all lines under 80 if possible
- Looks like some files in R/ don't have any tests yet, eventually get those tests in place.

```r
── GP photosearcher ─────
It is good practice to

  ✖ write unit tests for all functions, and all package code in general. 68% of code lines are covered by test
    cases.

    R/download_images.R:81:NA
    R/download_images.R:82:NA
    R/download_images.R:83:NA
    R/download_images.R:136:NA
    R/download_images.R:137:NA
    ... and 121 more lines

  ✖ avoid long code lines, it is bad for readability. Also, many people prefer editor windows that are about
    80 characters wide. Try make your lines shorter than 80 characters

    R/photo_search.R:108:1
    R/photo_search.R:180:1
    R/photo_search.R:184:1
    R/photo_search.R:190:1
    R/utils.R:32:1
    ... and 1 more lines

───────────
```

Seeking reviewers now 🕐

---

Reviewers:

- xx
- xx
