### 2019-08-07 dfc22ea
- added a better error check
