### 2019-08-07 21dc9c0
- Reverted mindate deafult
