### 2019-08-09 84d38c9
- updated relevant functions to include new error search
-fixed 80 character line limit
