+++ R/related_tags.R
@@ -1,40 +1,42 @@
-#' Get related terms
+#' Get related tags
 #'
-#' Takes a term and returns the top tags related to that term.
+#' Takes a tag and returns the top tags related to that tag.
 #'
 #' Uses the flickr.tags.getRelated API method from the Flickr API. See
 #' \url{https://www.flickr.com/services/api/flickr.tags.getRelated.html} for
 #' more information on the API method.
 #'
-#' @param term character. Term to search.
+#' @param tag character. tag to search.
 #'
-#' @return character. Tags most associated with input term.
+#' @return character. Tags most associated with input tag.
 #' @export
 #'
 #' @examples
 #' \dontrun{
-#' related_terms(term = "car")
+#' related_tags(tag = "car")
 #'
-#' related_terms(term = "monkey")
+#' related_tags(tag = "monkey")
 #'
-#' related_terms(term = "river")
+#' related_tags(tag = "river")
 #' }
-related_terms <- function(term) {
+related_tags <- function(tag) {
 
   # this checks for the presence of a key, if no key it prompts the user to
   # create one, it then checks the validity of the key
   api_key <- create_and_check_key()
 
-  term <- gsub(" ", "+", trimws(term))
+  tag <- gsub(" ", "+", trimws(tag))
 
   baseURL <- paste("https://api.flickr.com/services/rest/",
                    "?method=flickr.tags.getRelated&api_key=",
                    api_key,
                    "&tag=",
-                   term,
+                   tag,
                    sep = "")
 
   tag_xml <- search_url(base_url = baseURL)
+
+  find_errors(error_xml = tag_xml)
 
   if (!is.null(tag_xml)) {
     tag_atts <- xml2::xml_find_all(tag_xml, "//tag", ns = xml2::xml_ns(tag_xml))
@@ -42,7 +44,7 @@ related_terms <- function(term) {
   }
 
   if (is.null(tags)) {
-    stop("No related terms")
+    stop("No related tags")
   }
 
   return(tags)
