+++ tests/testthat/test-related_tags.R
@@ -1,19 +1,19 @@
-context("test-related_terms")
+context("test-related_tags")
 
 test_that("output is correct", {
   skip_on_cran()
   write.table("6a2ac025703c4b98aae141842eae8b1d", file = "api_key.txt")
-  term_test <- related_terms(term = "tree")
+  tag_test <- related_tags(tag = "tree")
 
-  expect_is(term_test, "character")
+  expect_is(tag_test, "character")
 })
 
 test_that("invalid API keys fails correctly", {
   # if(file.exists("api_key.txt")) { file.remove("api_key.txt") }
-  # expect_error(related_terms(term = "tree"), "Enter API key or save using the
+  # expect_error(related_tags(tag = "tree"), "Enter API key or save using the
   # save_key function")
 
   skip_on_cran()
   write.table("notarealkey", file = "api_key.txt")
-  expect_error(related_terms(term = "tree"))
+  expect_error(related_tags(tag = "tree"))
 })

[…]
