### 2019-08-09 #325 @nfox29
Thanks @sckott to address your comments I have fixed the line length issues. The reason currently some lines aren't covered by tests is due to a Flickr side issue. Currently anything that requires a Flickr location identifier does not work. Therefore both the `location_tags` and `find_place` functions do not work, as well as searching for photographs with a woeID using the `photo_search` function. Because of this any test that are related to those functions are currently being skipped until Flickr resolves the issue. More generally there are also a few lines of code that catch errors that I can not replicated in tests.

Kindest,
Nathan
