### 2019-08-12 66d3f07
- added interesting list function and associated tests and docs
