+++ NAMESPACE
@@ -4,6 +4,7 @@ export(download_images)
 export(find_place)
 export(get_exif)
 export(get_photoinfo)
+export(interesting_list)
 export(location_tags)
 export(photo_search)
 export(related_tags)
