### 2019-08-12 acf19f3
- minor changes to code and added additional tests
- test coverage up to 86.50
