+++ tests/testthat/test-download_images.R
@@ -34,6 +34,38 @@ test_that("height and width can be chosen", {
 
 })
 
+test_that("if no photos match height and width, skip", {
+  skip_on_cran()
+  write.table("6a2ac025703c4b98aae141842eae8b1d", file = "api_key.txt")
+  size_test <- download_images(photo_id = 47259127482,
+                               saveDir = "test_images",
+                               max_image_height = 1,
+                               max_image_width = 1)
+
+  expect_is(size_test, "data.frame")
+  expect_equal(ncol(size_test), 2)
+})
+
+test_that("if photo is not found, skip", {
+  skip_on_cran()
+  write.table("6a2ac025703c4b98aae141842eae8b1d", file = "api_key.txt")
+  notreal_test <- download_images(photo_id = "not a real photo")
+
+  expect_is(notreal_test, "data.frame")
+  expect_equal(ncol(notreal_test), 2)
+
+})
+
+test_that("if photo has no permission, skip", {
+  skip_on_cran()
+  write.table("6a2ac025703c4b98aae141842eae8b1d", file = "api_key.txt")
+  notreal_test <- download_images(photo_id = "1231231")
+
+  expect_is(notreal_test, "data.frame")
+  expect_equal(ncol(notreal_test), 2)
+
+})
+
 
 test_that("invalid API keys fails correctly", {
 
