+++ tests/testthat/test-get_photoinfo.R
@@ -4,7 +4,7 @@ test_that("output is correct", {
   skip_on_cran()
 
   write.table("6a2ac025703c4b98aae141842eae8b1d", file = "api_key.txt")
-  photo_test <- get_exif(photo_id = 47259127482)
+  photo_test <- get_photoinfo(photo_id = 47259127482)
 
   expect_is(photo_test, "data.frame")
 })
