+++ DESCRIPTION
@@ -9,7 +9,7 @@ Authors@R:
            person("Tom", "August", role = c("aut")))
 Description: Queries the Flick API (https://www.flickr.com/services/api/)to return photograph metadata as well as the ability to download the images as jpegs.
 Depends: R (>= 3.5.0)
-License: GPL
+License: GPL-3
 Encoding: UTF-8
 LazyData: true
 Language: en_GB
