### 2019-08-13 9389545
- edited user_info function
