### 2019-08-30 #325 @monicagerber
Hi @nfox29! 👋 I have a question about getting started with your package. In the README, you write:

>The package requires a valid API key from the Flickr development page. The first time you call a function from the package, you will be prompted to create and enter your API key. The API key will then be saved as a .txt file and be called when using any other function.

Do you have an example of how the user will be prompted to create and enter an API key? I've installed your package and tried out `photo_search()` but wasn't prompted to create a key. I went ahead and created an API key from the Flickr development page, but as a naïve user, I'm not sure where to save the key.
