### 2019-08-31 #325 @nfox29
@monicagerber when you use the `photo_search` function (or any of the other packages functions) you should be prompted to paste or type your API key into the console. All functions check whether an API key exists as a a file called `api_key.txt` and if not it should prompt you to enter it, whilst opening the Flickr development page to facilitate requesting an API key if you do not have one. It is then saved to be called upon for future functions. If an API key has been entered but is not a valid API key it will warn you that the API key is not valid and stop the function. Users are also able to save their API key directly as a file `api_tey.txt`. Let me know if you have continued issues with this and I will look into why it has not worked for you. Kindest regards.
