### 2019-09-02 #325 @zkamvar
---

My strategy for reading through this was to do the following:

1. work through the example in the README
2. attempt to adapt the example to my own interest
3. work through the examples in the vignette presented

I have installed the repository from the most recent commit, which was here:

```r
repo   <- "nfox29/photosearcher"
commit <- "@93895459fbde7a50b87ce0717a75e383ffc245dd"
remotes::install_github(paste0(repo, commit), dep = TRUE)
```

Overall, I think this is a good start, but it still needs some polishing. The
main function of this package seems to be `photo_search()`, which works well
once you have your API key set up (difficulty: not very). I was able to search
for snakes in the UK and got them in abundance, so the package works with that
expectation.

Based on my walkthrough of the documentation I think the entrypoints for the
API make sense, but the output could use significant improvement.

For example, the output of `photo_search()` has 57 columns, but the
documentation does not state exactly what they represent. I could tell that 30
of these columns were dedicated to different iterations of width, height, and
URL for the results with cryptic suffixes.  Moreover (as mentioned later), all
of the columns were encoded as character despite the fact that there were
integer, numeric, date, and boolean columns.

Moreover, the `get_photoinfo()` returns similar columns, but the `dateuploaded`
column is not formatted as the date columns of `photo_search()` (Y-m-d H:M:S),
but as an integer formatted as a character. The only other function that returns
a data frame `user_info()` converts characters to factors.

The other functions all return character vectors with the exception of
`download_images()`, which returns a data frame.

My recommendations for improvement:

1. Document the columns of `photo_search()`
2. Parse the data from `photo_search()`, `user_info()`, and `get_photoinfo()`
   before returning it to the user
3. Remove the default directory destination from `download_images()` and throw
   an error if the user doesn't supply one (this is one of the CRAN policies).
4. Make sure the vignette documentation works, even if you are not going to run
   it on CRAN.

Stray observations:

 - The examples documented in `photo_search()` have a cascading indentation
   pattern and will fill the user's console with verbose data frame output.
   Moreover, they might benefit from some comments describing what they are
   doing
 - The examples for `get_exif()` and `get_photoinfo()` do not work
 - "moutains" is a misspelling in the getting started vignette
 - The foxes photo is missing in the getting started vignette
 - the `api_key.txt` is a bit ambiguous a name for the key. I'm wondering if it
   would be better to use the package name and make it a hidden file along the
   lines of `.httr_oauth` (e.g. `.photosearcher_key`)
 - should the api key be kept secret or is it okay to be public?
 - It might be worthwhile investigating whether or not the `vcr` package can
   help with the testing coverage.

------------------------------------

#### Working through the README

The example in the README does not work out of the box:

``` r
library(photosearcher)

rock_climbing <- photo_search(
  mindate = "2010-01-01",
  maxdate = "2018-01-01",
  text = "rock climbing",
  bbox = "-12.875977,49.210420,2.636719,59.977005",
  has_geo = TRUE
)
#> Error in photo_search(mindate = "2010-01-01", maxdate = "2018-01-01", : argument 1 matches multiple formal arguments
```

I found that fixing `mindate` and `maxdate` to have `_taken` suffix, addressed
the problem. When I ran the command with the suffixes, I was prompted to find my
API key from flickr. I did not have one, so I used a throwaway account to
generate one. I copy+pasted the key (not the secret) into to console as it
prompted me to do so  and the search for eight years of rock climbers in the UK
and Ireland produced fruitful results, taking only a few seconds to search.

The example of plotting the geographic coordinates of all the rock
climbers in that eight year period was helpful and it worked as expected.

#### Adapting the README example

I wanted to see if I could show my spouse that yes, snakes did exist in the UK
despite the fact that we haven't seen one while hiking in the last year and was
able to produce a map of all photos tagged with "snake" in the last year.

When I tried to color the points by number of comments, I got a categorical
variable instead of a continuous because all the columns were encoded as
character. I found that this affected the SF object because it could not
interpret the character coordinates as a bounding box for the print method:

```
Simple feature collection with 148 features and 55 fields
geometry type:  POINT
dimension:      XY
bbox:           Error in signif(attr(x, "bbox"), options("digits")$digits) :
  non-numeric argument to mathematical function
```

Things worked well after I used the following pipeline to convert relevant
columns:

```r
library('dplyr')
snakes_clean <- snakes %>%
  mutate_at(vars(starts_with("count")), as.integer) %>%
  mutate_at(vars(ends_with("itude")), as.numeric) %>%
  mutate_at(vars(starts_with("height"), starts_with("width")), as.numeric) %>%
  mutate_at(vars(starts_with("date")), readr::parse_datetime)
```

#### Walk through the [package vignette](https://nfox29.github.io/photosearcher/articles/photosearcher.html)

The code from the vingette did not work out of the box for the following issues:

1. mindate/maxdate needed the `_taken` suffix
2. a comma was needed after `tags = "lake"`
3. no one took any pictures of a lake with the text "mountain" on New Year's 2019
4. I could not test the location services because they were down for flickr
5. the function `related_terms()` was renamed to `related_tags()`

I did not attempt the search by layer because I did not have an SF layer on hand.
I would recommend to make the layer you have in the tests available to the users
through something like `system.file()` so that they have a chance to run it if
they don't have a convenient layer to work through the example.

I think the `download_images()` demonstrataion needs to be modified a bit. If
the user naively walks through this and copy+pasting code, they will end up
downloading over a gigabyte of photos to a directory called "Fox images"
situated in their current working directory.

I would suggest the following change:

```r
# Downloading the first ten fox images from photosearch
temp_fox_dir <- dir.create(tempdir(), "foxes")
download_images(photo_id = foxes$id, saveDir = temp_fox_dir)

# Download a specific images based off its id
download_images(photo_id = 47416147181, saveDir = temp_fox_dir)

# View the images in the directory
if (interactive()) browseURL(temp_fox_dir)
```

**Important**: The `download_images()` function explicitly creates a directory
in the user's current working directory called "downloaded_images" or simply
attempts to write into this directory if this exists. It would be better to
throw an error if the user doesn't specify a directory.

The `related_tags()` function works as expected

The example for `get_photoinfo()` appears to be duplicated.
