### 2019-09-03 #325 @monicagerber
Hi @nfox29 and @sckott. I am nearly done with the review for the package but need until tomorrow to finish up. I got caught up with the Flickr API key. I realized that the issue was that the in the example in the README, the argument names should be `mindate_taken` and `maxdate_taken` instead of `mindate` and `maxdate`. The API key functionality works for me.
