### 2019-09-04 #325 @monicagerber
---

**A statement of need**

* Suggested edit for clarity:

> The package also supports a number of other Flickr API call methods including the flickr.tags.getRelated and flickr.places.tagsForPlace methods.

* Consider adding a brief (1 sentence) overview of the analyis in your Vignette ("Flickr in research") to your statement of need.

**Vignettes**

*Main vignette:*

* `mindate` and `maxdate` arguments do not work.
* Plot of UK foxes isn't visable.
* There are a lot of fox images in the example. It might be helpful to have an example with less images or download fewer of them in your example.
* `download_images(photo_id = 47416147181)`: Not able to download.
* `user <- user_info(user_id = "155421853@N05")`: Not a great example because the hometown is empty.
* Some typos in the vignette writing. I'm happy to review again once you've made other changes and check specifically for grammar/spelling. I can open a PR with suggested changes.
* Overall, you may want to provide a better narrative for why you might want to use the features of your package and how they all fit together.

*Article:*

* Would be helpful if there was more text here describing your motivation, results, and conclusion.

**Function documentation**

* Functions are documented except for `interesting_list()` (not on package website and documentation seems incomplete), `findPlaces()`, and `location_tags()`.
* `related_terms()` is not exported, but `related_tags()` is. I think the `related_tags()` name is more informative, but can you give more info on how relation is determined? Is it relation to tags or other text associated with image?
* Consider improving clarity in `@return` for `photo_search()` function and others. Link to data dictionary for data.frames with `@seealso`.

> data.frame. Output consists of 57 variables including; latitude and longitude of photograph, date and time it was taken, associated tags and image urls.

> A data.frame with 57 columns including latitude and longitude of the photograph, the date and time it was taking, and the associated tags and image urls.

* Functions download_images(), get_exif(), get_photoinfo() need a short informative title (like other funtions)
* `exif()`: Can you link to more information about what exif means? How is this different than `get_photoinfo()`?

**Examples**

* `download_images()`:
    + First example: Can you use a real photo_id?
    + Second example: I think the example should be self-contained. The user might not know what photo_search_outputs means.
    + Third example: Extra comma at end of example.
* `find_place()`, `location_tags()`: unable to test because Flickr locations services are down
* `get_exif()`, `get_photoinfo()`: Should use real photo_id
* `photo_search()`: code style/indentation could be improved

**Community guidelines**

* To be consistent with other packages, filename should be in CAPS.

* There are a couple of functions that use locations services that I wasn't able to test.

**Performance**

* Suggested enhancement: It takes a while to download images. Is there a way to warn the user of how long this might take?

**Packaging guidelines**

* I think it would be helpful to rename the functions in this package using the object_verb() naming scheme (see [section 1.4: Function and argument naming](https://devguide.ropensci.org/building.html#function-and-argument-naming)).
* It would also be nice to group functions into families (see section 1.8.1, "Grouping functions in the reference," and resulting reference section example). I think this would be helpful to me as a new user of the Flickr API. For example, you could group functions by place-finding functions (find_place, location_tags), getting meta data for known photos (get_exif, get_photoinfo, photo_search, related_terms, user_info), and finding photos (photo_search, download_images).

**Other comments**

* `photo_search()`:
    + Rather than listing the column names, could you add a link to the data dictionary in the details section? You could use the `@seealso` Roxygen tag.
    + Works if mindate and maxdate are in a Date format too.
    + If you search for "climbing rock" instead of "rock climbing" there are many less results. Can you point the user to more information on search strategies on Flickr? For example, can you use the `text` argument with Boolean logic? For example, how would you do a search for "rock climbing" but not include results with "trad"? Or "rock climbing" AND "sport"?
    + `tags`: Be more specific, do you mean a character vector?
    + Since `find_place()` and `woe_id()` are related functions, could add link to them where you mention them under `woe_id()` I think you can do this with `\code{\link{functioname}}`, see chapter 5 in R packages book.
    + What is an NSID?
    + Enhancement: could you check to see which photos can be downloaded?
* `download_images()`:
    + `photo_id`: Could be more specific about class. A numeric vector? A character vector also works.
    + What happens if images already exist in that directory? No warning is given.
    + There is a mix of camel and snake case for argument names.
* DESCRIPTION: Description goes over 80 characters -- can you make a new line?
