### 2019-09-05 #325 @nfox29
@monicagerber @zkamvar thank you both for your reviews, they are extremely helpful. Apologises for the errors in the examples that was an oversight on my behalf. I will now work through your suggested changes and will reply once I have finished. Kindest regards, Nathan.
