### 2019-09-09 41cbe58
- started edditing for r OpenSci submission - edits to vignettes and examples
