+++ vignettes/photosearcher.Rmd
@@ -12,55 +12,58 @@ vignette: >
 This vignette provides an overview of the functions in R package `photosearcher`.
 
 ## Searching for photographs
-These are a few basic examples of the main `photo_search()` function:
+The `photo_search()` function can be used to search Flickr for photographs that meet specific search criteria. For example, you could search for sightings of animals to better understand the species distribution:
 
 ```{r eval=FALSE}
 #Search for photos of foxes in the UK for the year 2017
-foxes <- photo_search(mindate = "2017-01-01",
-                      maxdate = "2018-01-01",
+foxes <- photo_search(mindate_taken = "2017-01-01",
+                      maxdate_taken = "2018-01-01",
                       text = "foxes",
                       bbox = "-12.875977,49.210420,2.636719,59.977005",
                       has_geo = TRUE)  
 
 #Search for images of trees globally for the 1st of January 2019
-trees <- photo_search(mindate = "2019-01-01",
-                      maxdate = "2019-01-01",
+trees <- photo_search(mindate_taken = "2019-01-01",
+                      maxdate_taken = "2019-01-01",
                       text = "tree",
                       has_geo = TRUE)
 
 #Search for images with text mountain and tagged lake
-moutain_lake <- photo_search(mindate = "2019-01-01",
-                             maxdate = "2019-01-01",
+moutain_lake <- photo_search(mindate_taken = "2019-01-01",
+                             maxdate_taken = "2019-02-01",
                              text = "mountain",
-                             tags = "lake"
+                             tags = "lake",
                              has_geo = FALSE)
 ```
 
-Users can also use the `sf` package to search via a simple features layer.
+Users can also use the `sf` package to search via a simple features layer, this can be used to find images of in a specific location for example to aid in resource management for recreational activites in english national parks.
 
 ```{r eval=FALSE}
 #load shape file
-sf_layer <- sf::read_sf(dsn = "path to layer")
+national_parks <- sf::read_sf(dsn = ".\\data\\shape_file\\National_Parks_England.shp")
 
 #Search for photos of foxes in the UK for the year 2017
-search_by_layer <- photo_search(sf_layer = sf_layer,
-                      has_geo = TRUE) 
+parks_hiking <- photo_search(mindate_taken = "2012-01-01",
+                             maxdate_taken = "2012-04-01",
+                             text = "hiking",
+                             sf_layer = national_parks,
+                             has_geo = TRUE) 
 ```
 
 When `has_geo == TRUE` only metadata about images with latitude and longitude information will be retrieved. 
 
 Here, the images of UK foxes are plotted using `ggplot2`. 
 
-![Foxes](https://github.com/nfox29/photosearcher/blob/master/images/foxes.png?raw=true)
+![Foxes](https://github.com/nfox29/photosearcher/data/images/foxes_plot.png?raw=true)
 
 ## Downloading images
-The package can also download the images based of of their unique id number. You can direct the download to any save directory, if it does not exists it will be automatically created.
+The package can also download the images based of of their unique id number. You can direct the download to any save directory, if it does not exists it will be automatically created. Downloading images allows for validation of their contents. Here, we demonstrate how to download the first five images that were found using the foxes `photo_search` example, as well as for a single image.
 ```{r eval=FALSE}
 #Downloading fox images from photosearch
-download_images(photo_id = foxes$id, saveDir = "Fox images")
+download_images(photo_id = foxes$id[1:5], saveDir = "Fox images")
 
 #Download a specific images based off its id
-download_images(photo_id = 47416147181)
+download_images(photo_id = 48704764812)
 ```
[…]
