### 2019-09-10 94e40fc
- updated function descriptions and website
