+++ R/interesting_list.R
@@ -1,21 +1,24 @@
 #' interesting_list
 #'
+#' Returns a Flickr generated list of photographs deemed interesting. For
+#' information on how this list is calculated see:
+#' \link{http://www.steves-digicams.com/knowledge-center/how-tos/online-sharing-social-networking/what-is-flickr-interestingness.html#b}
+#'
 #' @param date Character, required. Interestingness list for the date provided.
 #'   The date should be in the form of "YYYY-MM-DD".
 #'
 #' @returndata.frame. Output consists of 57 variables including; latitude and
-#'   longitude of photograph, date and time it was taken, associated tags and
-#'   image urls.
+#' longitude of photograph, date and time it was taken, associated tags and
+#' image urls.
 #'
-#'   Full list of variables returned: id, owner, secret, server, farm, title,
-#'   ispublic, isfriend, isfamily, license, datetaken, datetakengranularity,
-#'   datetakenunknown, count_views, count_comments, count_faves, tags, latitude,
-#'   longitude, accuracy, context, place_id, woeid, geo_is_family,
-#'   geo_is_friend, geo_is_contact, geo_is_public, url_sq, height_sq, width_sq,
-#'   url_t, height_t, width_t, url_s, height_s,	width_s	url_q, height_q,
-#'   width_q, url_m, height_m, width_m, url_n, height_n, width_n, url_z,
-#'   height_z, width_z, url_c, height_c, width_c, url_l, height_l, width_l,
-#'   url_o, height_o, width_o.
+#' Full list of variables returned: id, owner, secret, server, farm, title,
+#' ispublic, isfriend, isfamily, license, datetaken, datetakengranularity,
+#' datetakenunknown, count_views, count_comments, count_faves, tags, latitude,
+#' longitude, accuracy, context, place_id, woeid, geo_is_family, geo_is_friend,
+#' geo_is_contact, geo_is_public, url_sq, height_sq, width_sq, url_t, height_t,
+#' width_t, url_s, height_s,	width_s	url_q, height_q, width_q, url_m, height_m,
+#' width_m, url_n, height_n, width_n, url_z, height_z, width_z, url_c, height_c,
+#' width_c, url_l, height_l, width_l, url_o, height_o, width_o.
 #'
 #' @export
 #'
