### 2019-09-10 149e3f5
- updated man and docs
