+++ R/utils.R
@@ -10,6 +10,7 @@ get_url <- function(mindate_taken,
                     page,
                     text = NULL,
                     tags = NULL,
+                    tag_mode = NULL,
                     bbox = NULL,
                     woe_id = NULL,
                     has_geo = TRUE) {
@@ -18,6 +19,7 @@ get_url <- function(mindate_taken,
                     "?method=flickr.photos.search&api_key=", api_key,
                     "&text=", text,
                     "&tags=", tags,
+                    "&tag_mode=", tag_mode,
                     "&min_taken_date=", as.character(mindate_taken),
                     "&max_taken_date=", as.character(maxdate_taken),
                     ifelse(!(is.null(mindate_uploaded)), paste0(
