### 2019-09-11 26d6e7a
- updated documentation for phot_search and interesting_list
- added test for tag_all in photo_search
