### 2019-09-11 07a2996
- Fixed failing tests
