### 2019-09-11 b2597dc
- now parses user_info and get_photoinfo outputs
