### 2019-09-12 8d0cd1f
update the output of photosearch to bulletpoints and capitlise contributing.md
