+++ man/related_tags.Rd
@@ -29,3 +29,4 @@ related_tags(tag = "monkey")
 related_tags(tag = "river")
 }
 }
+\concept{Tag information}
