+++ R/utils.R
@@ -110,7 +110,7 @@ ui_info <- function(x, .envir = parent.frame()) {
 # this checks for the presence of a key, if no key it prompts the user to create
 # one, it then checks the validity of the key
 create_and_check_key <- function() {
-  if (!file.exists("api_key.txt")) {
+  if (!file.exists("photosearcher_key.txt")) {
     ui_todo(
       "Create a Flickr API key at https://www.flickr.com/services/apps/create/")
 
@@ -119,12 +119,13 @@ create_and_check_key <- function() {
     ui_todo("Enter your Flickr API key:")
 
     utils::write.table(readline(),
-                       file = "api_key.txt",
+                       file = "photosearcher_key.txt",
                        col.names = FALSE,
                        row.names = FALSE)
   }
 
-  api_key <- utils::read.table("api_key.txt", stringsAsFactors = FALSE)
+  api_key <- utils::read.table("photosearcher_key.txt",
+                               stringsAsFactors = FALSE)
 
   base_url <- paste("https://api.flickr.com/services/rest/",
                     "?method=flickr.photos.search&api_key=",
@@ -136,7 +137,7 @@ create_and_check_key <- function() {
   warn <- as.character(unlist(pages_data))
 
   if ((warn[2]) == ("Invalid API Key (Key has invalid format)")) {
-    stop("Invalid API Key: correct this in api_key.txt")
+    stop("Invalid API Key: correct this in photosearcher_key.txt")
   }
 
   return(api_key)
@@ -195,24 +196,9 @@ parse_pic <- function(pics = NULL){
   cols.num <- c("id",
                 "server",
                 "farm",
-                "ispublic",
-                "isfriend",
-                "isfamily",
-                "license",
-                "datetakengranularity",
-                "datetakenunknown",
-                "count_views",
-                "count_faves",
-                "count_comments",
                 "latitude",
                 "longitude",
-                "accuracy",
-                "context",
                 "woeid",
-                "geo_is_family",
-                "geo_is_friend",
-                "geo_is_contact",
-                "geo_is_public",
                 "height_sq",
                 "width_sq",
                 "height_t",
@@ -235,6 +221,24 @@ parse_pic <- function(pics = NULL){
                 "width_o")
 
   pics[cols.num] <- sapply(pics[cols.num],as.numeric)
+
+  cols.num <- c("ispublic",
+                "isfriend",
+                "isfamily",
+                "license",
+                "datetakengranularity",
+                "datetakenunknown",
+                "count_views",
+                "count_faves",
+                "count_comments",
+                "accuracy",
+                "context",
+                "geo_is_family",
+                "geo_is_friend",
+                "geo_is_contact",
+                "geo_is_public")
+
+  pics[cols.num] <- sapply(pics[cols.num],as.integer)
 
   return(pics)
 
