### 2019-09-16 7669240
- download images now requires a file directory to be provided
- download images now has an argument for whether user want files to be overwritten. If true user will also be warned at the end which files were overwritten

- added new tests for download_images

- added working shape file for vingette example
