### 2019-09-16 9b54645
- fixed failing tests
