### 2019-09-16 8bc4fc9
- updated docs and fixed spelling mistakes - all examples and vignettes now run
