+++ R/photo_search.R
@@ -1,120 +1,112 @@
-#' Search for photo metadata
+#'Search for photo metadata
 #'
-#' Returns image metadata for photos matching the search terms.
+#'Returns image metadata for photos matching the search terms.
 #'
-#' Uses the flickr.photos.search API method from the Flickr API. This search
-#' method requires a limiting factor to prevent parameterless searches - to
-#' ensure this is met the function requires both a minimum and a maximum date
-#' that searched photographs were taken on. See
-#' \url{https://www.flickr.com/services/api/flickr.photos.search.html} for more
-#' information on the API method.
+#'Uses the flickr.photos.search API method from the Flickr API. This search
+#'method requires a limiting factor to prevent parameterless searches - to
+#'ensure this is met the function requires both a minimum and a maximum date
+#'that searched photographs were taken on. See
+#'\url{https://www.flickr.com/services/api/flickr.photos.search.html} for more
+#'information on the API method.
 #'
-#' @param mindate_taken Character, or date required. Minimum taken date. Photos
-#'   with an taken date greater than or equal to this value will be returned.
-#'   The date should be in the form of "YYYY-MM-DD".
-#' @param maxdate_taken Character, or date required. Maximum taken date. Photos
-#'   with an taken date less than or equal to this value will be returned. The
-#'   date should be in the form of "YYYY-MM-DD".
-#' @param text Character, optional. A free text search. Photos who's title,
-#'   description or tags contain the text will be returned. You can exclude
-#'   results that match a term by prepending it with a - character. Free text
-#'   searches for words in order provided, for example a search for "climbing
-#'   rock" will be different to "rock climbing".
-#' @param tags Character vector, optional. A comma-delimited list of tags.
-#'   Photos with one or more of the tags listed will be returned. You can
-#'   exclude results that match a term by prepending it with a - character.
-#' @param tags_any Logical, optional. If TRUE, photos containing any of the tags
-#'   will be returned. If FALSE, only photos containg all tags will be returned.
-#'   Defulted to return any tags.
-#' @param bbox String, optional bounding box of search area provide as:
-#'   "minimum_longitude,minimum_latitude,maximum_longitude,maximum_latitude".
-#' @param woe_id Numeric, optional "where on earth identifier" can be supplied
-#'   instead of bbox. Use \code{\link{find_place}} to obtain woe_id for a place.
-#' @param sf_layer A simple features layer, optional, area to search for photos,
-#'   can be supplied instead of a bbox or woeID.
-#' @param has_geo Logical, optional argument for whether returned photos need
-#'   associated spatial data.
-#' @param mindate_uploaded Character or date, optional. Minimum upload date.
-#'   Photos with an upload date greater than or equal to this value will be
-#'   returned. The date can be in the form of a unix timestamp or mysql
-#'   datetime.
-#' @param maxdate_uploaded Character or date, optional. Maximum upload date.
-#'   Photos with an upload date less than or equal to this value will be
-#'   returned. The date can be in the form of a unix timestamp or mysql
-#'   datetime.
-#' @param user_id Character, optional. The Flickr ID of the user who's photo to
-#'   search. If this parameter isn't passed then everybody's public photos will
-#'   be searched.
+#'@param mindate_taken Character, or date required. Minimum taken date. Photos
+#'  with an taken date greater than or equal to this value will be returned. The
+#'  date should be in the form of "YYYY-MM-DD".
+#'@param maxdate_taken Character, or date required. Maximum taken date. Photos
+#'  with an taken date less than or equal to this value will be returned. The
+#'  date should be in the form of "YYYY-MM-DD".
[…]
