+++ R/user_info.R
@@ -19,8 +19,6 @@
 #' @examples
 #' \dontrun{
 #' user_info(user_id = "155421853@N05")
-#'
-#' user_info(user_id = photo_search_outputs$owner)
 #' }
 user_info <- function(user_id) {
 
