### 2019-09-17 45c1ee7
- edits to docs
