### 2019-09-17 ca00903
- photo_info now returns correct date format
