### 2019-09-25 3e1feae
moved whitespace removal into get_url; removed iso encoding as causing issues with special characters
