### 2019-10-03 d298a55
- added package logo to read me
