### 2019-10-10 #325 @zkamvar
I am happy with the response. I think @nfox29 did a thorough job addressing our comments and it has improved the package experience.

I only have a couple of suggestions/questions for the documentation/behavior of `photo_search()`:

1. It's not clear that the user is expected to have a file called `photosearcher_key.sysdata` in their working directory, nor that it will be explicitly written to the directory when they enter their api key. It would be good to have this documented in the `photo_searcher()` function at the least.
2. If possible, it would be preferred to take the "name" element of the license attribute instead of the "id" element since that gives a description of the license
3. There are several variables in the table that are effectively logical, but are stored as integers (e.g. ispublic). I would recommend to store these as logical values so that it is explicit.
4. Should the ID (id, woeid) elements be stored as characters instead of integers since they are identifiers?
