### 2019-10-15 11b75a7
- added progress bar to photo_search
