### 2019-10-18 f3ffeb4
- changed the method of splitting smaller searchers - now returns photographs based on date taken ascended - takes the date of the last image taken and use this to search for it
