+++ R/photo_search.R
@@ -193,47 +193,9 @@ photo_search <-
 
     } else {
 
-        tags_any <- "all"
+      tags_any <- "all"
 
     }
-
-    #find number of results
-    # set search dates
-    mindate_taken <- as.POSIXct(date_df[1, "mindate_taken"])
-    maxdate_taken <- as.POSIXct(date_df[1, "maxdate_taken"])
-
-    # rest page to 1
-    i <- 1
-
-    base_url <- get_url(
-      mindate_taken = mindate_taken,
-      maxdate_taken = maxdate_taken,
-      mindate_uploaded = mindate_uploaded,
-      maxdate_uploaded = maxdate_uploaded,
-      user_id = user_id,
-      api_key = api_key,
-      page = i,
-      text = text,
-      tags = tags,
-      tag_mode = tags_any,
-      bbox = bbox,
-      woe_id = woe_id,
-      has_geo = has_geo
-    )
-
-    photo_xml <- search_url(base_url = base_url)
-
-    find_errors(error_xml = photo_xml)
-
-    if (!is.null(photo_xml)) {
-      pages_data <- data.frame(
-        xml2::xml_attrs(xml2::xml_children(photo_xml)))
-      pages_data[] <- lapply(
-        pages_data, FUN = function(x) as.integer(as.character(x)))
-      all_total <- pages_data["total", ]
-    }
-
-    pb <- txtProgressBar(min = 0, max = all_total, style = 3)
 
     # start while loop - until all dates are looped through
     while (nrow(date_df) > 0) {
@@ -242,113 +204,157 @@ photo_search <-
       mindate_taken <- as.POSIXct(date_df[1, "mindate_taken"])
       maxdate_taken <- as.POSIXct(date_df[1, "maxdate_taken"])
 
-      # rest page to 1
-      i <- 1
+      #flickr seems to search 8 hours in advance of
+      mindate_taken <- mindate_taken - 28800
 
-      base_url <- get_url(
-        mindate_taken = mindate_taken,
-        maxdate_taken = maxdate_taken,
-        mindate_uploaded = mindate_uploaded,
-        maxdate_uploaded = maxdate_uploaded,
-        user_id = user_id,
-        api_key = api_key,
-        page = i,
-        text = text,
-        tags = tags,
-        tag_mode = tags_any,
-        bbox = bbox,
-        woe_id = woe_id,
-        has_geo = has_geo
-      )
+      mindate_unix <- as.numeric(mindate_taken)
+      maxdate_unix <- as.numeric(maxdate_taken)
 
-      photo_xml <- search_url(base_url = base_url)
+      if (mindate_taken > maxdate_taken){
 
-      #add to number of needed calls
-      num_calls <- num_calls + 1
+        print("didn't get data")
 
-      find_errors(error_xml = photo_xml)
+        date_df <- date_df[-1, ]
 
-      if (!is.null(photo_xml)) {
-        pages_data <- data.frame(
-          xml2::xml_attrs(xml2::xml_children(photo_xml)))
-        pages_data[] <- lapply(
-          pages_data, FUN = function(x) as.integer(as.character(x)))
-        total_pages <- pages_data["pages", ]
-        total <- pages_data["total", ]
+      } else {
 
-        # if total > 4000 and dates are not 1 second apart, split
-        if (total > 4000 && (seq(
-          mindate_taken, length.out = 2, by = "1 secs")[2] != maxdate_taken)) {
-          x <- ceiling(total / 4000)
+        base_url <- get_url(
+          mindate_taken = mindate_unix,
+          maxdate_taken = maxdate_unix,
+          mindate_uploaded = mindate_uploaded,
+          maxdate_uploaded = maxdate_uploaded,
[…]
