### 2019-10-18 bff1e17
-removed print from photo_search
