### 2019-10-21 d6d823e
- edited tests that were taking too long
