### 2019-10-21 3a7e1a2
- some tests still taking a long time
