### 2019-10-29 cf3bdf1
- changed how results are clipped to shapefiles to speed up process
