### 2019-10-30 f8cd9e1
- Clarified the photosearcher_key.sysdata in all documents
- Changed the type of some dataframe outputs (i.e is.family is not logical)

- added license anem and url to outputs for clarity
