+++ R/location_tags.R
@@ -5,6 +5,12 @@
 #' \url{https://www.flickr.com/services/api/flickr.places.tagsForPlace.html} for
 #' more information on the API method.
 #'
+#'When running the function you need an API key saved as
+#'photosearch_key.sysdata in your working directory. If this is the first
+#'function you run you will be prompted to create and enter your API key. The
+#'API key will then be saved as photosearcher_key.sysdata in your working
+#'directory and is used for all function.
+#'
 #' @param woe_id numeric. a "Where on Earth" location tag.
 #'
 #' @return character. List of the top 100 tags associated with the woe_id.
