+++ R/user_info.R
@@ -9,6 +9,12 @@
 #' See \url{https://www.pixsy.com/academy/flickr-id/} for a guide on finding
 #' your Flickr ID.
 #'
+#'When running the function you need an API key saved as
+#'photosearch_key.sysdata in your working directory. If this is the first
+#'function you run you will be prompted to create and enter your API key. The
+#'API key will then be saved as photosearcher_key.sysdata in your working
+#'directory and is used for all function.
+#'
 #' @param user_id character. The id of the user you wish to obtain information
 #'   for.
 #'
