+++ man/related_tags.Rd
@@ -19,6 +19,12 @@ Takes a tag and returns the top tags related to that tag.
 Uses the flickr.tags.getRelated API method from the Flickr API. See
 \url{https://www.flickr.com/services/api/flickr.tags.getRelated.html} for
 more information on the API method.
+
+When running the function you need an API key saved as
+photosearch_key.sysdata in your working directory. If this is the first
+function you run you will be prompted to create and enter your API key. The
+API key will then be saved as photosearcher_key.sysdata in your working
+directory and is used for all function.
 }
 \examples{
 \dontrun{

[…]
