### 2019-10-30 #325 @nfox29
Hi @zkamvar,

I have addressed all of your additional comments:
1. I have clarified the API key storage and use in all documents not just for the one function.
2. I have added additional two additional columns to the output: license name and the url of information on that license
3. I have changed these to logical and have also changed all the dates so they are in the same format
4. This has be fixed

@sckott @monicagerber is there any additional changes I need to make. I am currently in the process of finishing my manuscript for MEE and would like to send the manuscript off after the rOpenSci process has been completed.

Kindest regards,
Nathan
