+++ R/get_photoinfo.R
@@ -3,7 +3,7 @@
 #' Returns image metadata for a single photograph.
 #'
 #'When running the function you need an API key saved as
-#'photosearch_key.sysdata in your working directory. If this is the first
+#'photosearcher_key.sysdata in your working directory. If this is the first
 #'function you run you will be prompted to create and enter your API key. The
 #'API key will then be saved as photosearcher_key.sysdata in your working
 #'directory and is used for all function.
