+++ R/photo_search.R
@@ -10,7 +10,7 @@
 #'information on the API method.
 #'
 #'When running the function you need an API key saved as
-#'photosearch_key.sysdata in your working directory. If this is the first
+#'photosearcher_key.sysdata in your working directory. If this is the first
 #'function you run you will be prompted to create and enter your API key. The
 #'API key will then be saved as photosearcher_key.sysdata in your working
 #'directory and is used for all function.
@@ -148,8 +148,8 @@
 #'
 #' }
 photo_search <-
-  function(mindate_taken = "2019-01-01",
-           maxdate_taken = "2019-01-01",
+  function(mindate_taken = NULL,
+           maxdate_taken = NULL,
            mindate_uploaded = NULL,
            maxdate_uploaded = NULL,
            user_id = NULL,
