### 2019-10-30 #325 @zkamvar
LGTM. Thank you for addressing those, @nfox29!
