### 2019-10-30 f2a0f3f
- edited failing tests
