+++ tests/testthat/test-photo_search.R
@@ -9,7 +9,9 @@ test_that("fails correctly", {
 
   skip_on_cran()
 
-  expect_error(photo_search(bbox = "324134,12341341,123413241,312412"),
+  expect_error(photo_search(bbox = "123123123,123,1231231,123",
+                            mindate_taken = "2019-01-01",
+                            maxdate_taken = "2019-02-01"),
                "Not a valid bounding box")
 
   #whilst flickr services are down
@@ -18,9 +20,12 @@ test_that("fails correctly", {
 
   #no photographs meet criteria
   expect_error(photo_search(tags = c("big", "dog", "mad", "ship", "hot", "old"),
-                            tags_any = FALSE))
+               tags_any = FALSE),
+               mindate_taken = "2019-01-01",
+               maxdate_taken = "2019-02-01")
 
   expect_error(photo_search(mindate_taken = "2017-01-01",
+                            maxdate_taken = "2017-01-01",
                             user_id = "155421853@N05"))
 
 })
@@ -29,7 +34,10 @@ test_that("bbox + woe_id fails correctly", {
   write.table("6a2ac025703c4b98aae141842eae8b1d",
               file = "photosearcher_key.sysdata")
   expect_error(
-    photo_search(bbox = "-7.86,54.62,-1.0,58.83", woe_id = 12578048),
+    photo_search(bbox = "-7.86,54.62,-1.0,58.83",
+                 woe_id = 12578048,
+                 mindate_taken = "2019-01-01",
+                 maxdate_taken = "2019-02-01"),
     "Specify search location as only one of: woe_id, bbox or sf_layer."
   )
 })
@@ -39,10 +47,11 @@ test_that("output is correct", {
   write.table("6a2ac025703c4b98aae141842eae8b1d",
               file = "photosearcher_key.sysdata")
 
-  tree_test <- photo_search(text = "tree",
-                            maxdate_taken = "2019-01-02")
-  expect_is(tree_test, "data.frame")
-  expect_equal(ncol(tree_test), 61)
+  text_test <- photo_search(text = "tree",
+                           mindate_taken = "2019-01-01",
+                           maxdate_taken = "2019-01-02")
+  expect_is(text_test, "data.frame")
+  expect_equal(ncol(text_test), 61)
 
   bbox_test <- photo_search(
     bbox = "-140.625000,-47.517201,167.695313,69.162558",
@@ -53,8 +62,10 @@ test_that("output is correct", {
 
 
   date_test <- photo_search(mindate_uploaded = "2019-01-01",
-                            maxdate_uploaded = "2019-02-01",
-                            text = "lake")
+                            maxdate_uploaded = "2019-03-01",
+                            mindate_taken = "2019-01-01",
+                            maxdate_taken = "2019-01-02",
+                            text = "tree")
   expect_is(date_test, "data.frame")
   expect_equal(ncol(date_test), 61)
 
@@ -70,12 +81,13 @@ test_that("shape files work", {
   write.table("6a2ac025703c4b98aae141842eae8b1d",
               file = "photosearcher_key.sysdata")
 
-  national_parks = sf::st_read(system.file("shape/National_Parks_England.shp",
+  national_parks <- sf::st_read(system.file("shape/National_Parks_England.shp",
                                            package="photosearcher"))
 
-  expect_warning(shape_test <- photo_search(mindate_taken = "2018-12-20",
-                             sf_layer = national_parks,
-                             text = "walk"))
+  expect_warning(shape_test <- photo_search(mindate_taken = "2018-12-28",
+                                            maxdate_taken = "2019-01-01",
+                                            sf_layer = national_parks,
+                                            text = "walk"))
 
   expect_is(shape_test, "data.frame")
   expect_equal(ncol(shape_test), 63)
