### 2019-10-30 #325 @sckott
Approved (with some caveats, see below)! Thanks @nfox29 for submitting and @zkamvar & @monicagerber for your reviews!

caveats: on check I see:

- On `checking Rd cross-references`, check is throwing warnings about missing links, are you not getting that?
- Getting `Files named as vignettes but with no recognized vignette engine` note about the `vignettes/flickr_in_research.Rmd` file. shouldn't that have a vignette engine?

- [ ] Add a `CODE_OF_CONDUCT.md` file, see `usethis::use_code_of_conduct`
- [ ] I'd suggest adding issue and pull request template files, see https://github.com/ropensci/dotgithubfiles
"```
- [ ] change your `pkgdown` URL to point to `https://docs.ropensci.org/package_name` and deactivate the automatic deployment you might have set up, since it will not be built centrally like for all rOpenSci packages, see http://devdevguide.netlify.com/#docsropensci. In addition, in your DESCRIPTION file, include the docs link in the `URL` field alongside the link to the GitHub repository, e.g.: `URL: https://docs.ropensci.org/foobar (website) https://github.com/ropensci/foobar`
- [ ] Add a mention of the review in `DESCRIPTION` via [`rodev::add_ro_desc()`](https://docs.ropensci.org/rodev/reference/add_ro_desc.html).
