### 2019-11-20 1c56773
- fixed vingette issues - fixed missing link issues
