+++ R/get_exif.R
@@ -2,7 +2,7 @@
 #'
 #'Returns Exchangeable image file format data for a single photograph. For more
 #'information on how Exif data differs from metadata see:
-#'\link{https://www.flickr.com/services/api/flickr.photos.getExif.html}
+#'\url{https://www.flickr.com/services/api/flickr.photos.getExif.html}
 #'
 #'When running the function you need an API key saved as
 #'photosearcher_key.sysdata in your working directory. If this is the first
