### 2019-11-20 7761bca
- added code of conduct, issue and puull request templates
