+++ _pkgdown.yml
@@ -1,1 +1,2 @@
 destination: docs
+url: https://docs.ropensci.org/photosearcher
