+++ DESCRIPTION
@@ -26,8 +26,9 @@ Imports:
     sf
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 6.1.1
-URL: https://github.com/nfox29/photosearcher
-BugReports: https://github.com/nfox29/photosearcher/issues
+URL: https://docs.ropensci.org/photosearcher 
+(website) https://github.com/ropensci/photosearcher
+BugReports: https://github.com/ropensci/photosearcher/issues
 Suggests: 
     httptest,
     knitr,
