+++ DESCRIPTION
@@ -10,7 +10,7 @@ Authors@R:
            person("Tom", "August", role = c("aut")))
 Description: Queries the Flick API (https://www.flickr.com/services/api/) to 
   return photograph metadata as well as the ability to download the images as 
-  jpegs.
+  jpegs.This package has been peer-reviewed by rOpenSci (v. 1.0).
 Depends: R (>= 3.5.0)
 License: GPL-3
 Encoding: UTF-8
